requirements.txt
Cython>=0.28,<1.0; python_version == '3.7'
numpy==1.18.2
git+https://github.com/FragileTech/atari-py.git#egg=atari-py
opencv-python==4.2.0.32
pillow-simd==7.0.0.post3
gym==0.17.1
pytest==5.4.1
